+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- feat: Add partner parameter ([#837](https://github.com/getsentry/sentry-wizard/pull/837))
+
 ## 4.1.0
 
 - feat(nuxt): More granular error catching while modifying config ([#833](https://github.com/getsentry/sentry-wizard/pull/833))